test: add basic level of testing to project
This commit adds a basic level of test to the project. It is a script
that spins up two projects (identical) along with dcdc and then performs
some quick accessibility checks via curl and psql from outside and
internal to the docker network.

The dnsmasq-ext/external.host is updated to better reflect the default
state of the tests.